Refactor out all of state
